requirements.txt
flask
flask_bootstrap
flask_wtf
wtforms
wtforms.validators
flask_sqlalchemy
werkzeug
flask_login
datetime
flask_restful
requests
python-dotenv
Flask-HTTPAuth
uuid
PyJWT
passlib
pymongo
dnspython